Use package.el for package management
It used to be straight.el for managing packages, which is considered as
a modern way to replace the default package.el. One of mechanics of
straight.el is to download recipes from somewhere different than
package.el. This works well if the Internet is fully accessible.
However, in a stricted environment, this is not possible. Only a mirror
to elpa, the package.el place, is provided.

As a consequnce, we has changed the package manager from straight.el to
package.el.